// worktimetracker.client/src/generate-api/models/NotificationDto.ts
/**
 * WorkTimeTracker.Server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { NotificationType } from '../models/NotificationType';
import { HttpFile } from '../http/http';

export class NotificationDto {
    'id': number;
    'title': string;
    'message': string;
    'type': NotificationType;
    'isRead': boolean;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "title",
            "baseName": "title",
            "type": "string",
            "format": ""
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "string",
            "format": ""
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "NotificationType",
            "format": ""
        },
        {
            "name": "isRead",
            "baseName": "isRead",
            "type": "boolean",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return NotificationDto.attributeTypeMap;
    }

    public constructor() {
    }
}